[Claude] Extract standalone text cleaner from text_splitter
Standalone clean_text() function with zero align_data dependencies (only re +
regex). Extracted from the cleaning regexes in text_splitter.chunks() for use
by consumers that can't import the full package.
